requirements.txt
langchain
python-dotenv
# ipykernel
langchain-community
# pypdf
# bs4
# arxiv
# wikipedia
# langchain-text-splitters
# langchain-openai
# chromadb
# sentence_transformers
# langchain_huggingface
# faiss-cpu
# langchain_chroma
# duckdb
pandas
# openai
langchain-groq
transformers
# duckduckgo_search==5.3.1b1
# pymupdf
# mysql-connector-python
# SQLAlchemy
# validators==0.28.1
# youtube_transcript_api
# unstructured
# pytube
# numexpr
# huggingface_hub
# streamlit